Move processing's models into models.py
It didn't make sense to have two differente files for the same and
"models" is a more descriptive name.
